improvements to textml
6/10/24:
Bug fix: It was not using color options when "gradients are per line" was false.
Changed default setting of "gradients are per line" to true.
Improved text of option label to describe function of "Enable multi-line features" checkbox
Added code to deal with degenerate saturation situation which was making it appear like color effects were not working when font color was pure white.
Built for OBS 30.1